Define a DTO for the auction
The auction I want to transfer to the application is not excatly the same as
the entity
-> It has the id of the house instead of the house itself, and does not have the current price.

I decided not to set the current price because this property is exclusively related
to how an auction works. I don't see a good reason to let the client set this property
him/herself.