Set the current price of an auction as nullable
When an auction is registered, the current price is not defined.
So I use the null reference to define that.

Since the auction can be created by its DTO at registration,
the value of the current price in the DTO can put the auction in an inconsistaent state
(as a business object). That is why I set the current price to null at registration.