Delete data in cascade
When an auction house must be deleted, the application checks if
it can delete safely by checking if no auction is running.
If it is OK, it deletes the auctions. However, in order to delete those auctions,
the user bid related to them must be deleted first.

That is why this commit add this deletion.